infra/deploy/pyproject.toml
[project]
name = "frontend-deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.8",
    "python-magic>=0.4.27",
    "python-magic-bin>=0.4.14",
]
